+++ binjr/src/main/resources/views/ChartPropertiesView.fxml
@@ -67,8 +67,12 @@
                     </HBox>
                 </VBox>
                 <Separator prefWidth="200.0"/>
-                <CheckBox fx:id="autoScaleYAxis" mnemonicParsing="false" text="Auto Scale Y Axis"/>
-                <HBox alignment="CENTER_LEFT">
+                <!--<CheckBox fx:id="autoScaleYAxis" mnemonicParsing="false" text="Auto Scale Y Axis"/>-->
+                <Label text="Auto Scale Y Axis" wrapText="true"/>
+                <VBox fillWidth="false">
+                    <ToggleSwitch fx:id="autoScaleYAxis" prefWidth="1.0"/>
+                </VBox>
+                <HBox fx:id="yAxisScaleSettings" alignment="CENTER_LEFT">
                     <children>
                         <GridPane hgap="5.0" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" vgap="10.0">
                             <columnConstraints>
